This will be executed on new environments in order to populate the database with the full list of schools based on the Get Information about Schools service from the DFE.

Also adds the GIAS CSV to version control which is 62MB large and over the 50MB warning limit that GitHub issues. Their suggestion is to use `git-lfs` for managing these sort of files, but since they don't outright block files under 100MB, I think it's simpler to not use LFS right this moment since the CSV is unlikely to double in size any time soon.

TODO:

- [ ] Add it to the seeds script so it runs on new environments and Heroku
- [x] Use https://github.com/zdennis/activerecord-import for potentially much faster imports

### Usage

#### Without activerecord-import

```sh
mavis (gias-schools) $ rails db:seed:replant
mavis (gias-schools) $ time rails schools:import
Starting schools import. Total locations: 1
Progress: 51128/51128 [##################################################] 100.0%
Schools import completed. Total locations: 51128

________________________________________________________
Executed in  185.26 secs    fish           external
   usr time  124.03 secs    0.07 millis  124.03 secs
   sys time    6.69 secs    1.79 millis    6.69 secs
mavis (gias-schools) $ time rails schools:import
Starting schools import. Total locations: 51128
Progress: 51128/51128 [##################################################] 100.0%
Schools import completed. Total locations: 51128

________________________________________________________
Executed in   35.60 secs    fish           external
   usr time   29.13 secs    0.08 millis   29.13 secs
   sys time    1.53 secs    1.94 millis    1.53 secs
```

#### With activerecord-import

```
mavis (gias-schools) $ rails db:seed:replant
mavis (gias-schools) $ time rails schools:import
Starting schools import. Total locations: 1
Progress: 51128/51128 [##################################################] 100.0%
Schools import completed. Total locations: 51128

________________________________________________________
Executed in   16.91 secs    fish           external
   usr time   14.15 secs    0.09 millis   14.15 secs
   sys time    0.44 secs    1.79 millis    0.44 secs

mavis (gias-schools) $ time rails schools:import
Starting schools import. Total locations: 51128
Progress: 51128/51128 [##################################################] 100.0%
Schools import completed. Total locations: 51128

________________________________________________________
Executed in   17.29 secs    fish           external
   usr time   14.18 secs    0.10 millis   14.18 secs
   sys time    0.45 secs    1.05 millis    0.45 secs
```